前置推送重置标签算法

+++ graph/adj_graph.h
@@ -157,7 +157,7 @@ public:
     }
 
     //顶点u在最大流中的超额流
-    size_t &excess(size_t u) {
+    int &excess(size_t u) {
         assert(u < this->nodeNum && nodes[u]);
         return nodes[u]->e;
     }
